+++ dtor/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2026-05-03
+
+### Fixed
+
+- WASM support now uses `atexit` on all platforms (including
+  `wasm32-unknown-unknown`).
+- Typos and other documentation nits.
+
 ## [0.13.1] - 2026-05-02
 
 ### Changed